+++ core/class/Abeille.class.php
@@ -490,6 +490,7 @@ class Abeille extends eqLogic {
                 $zgType = $config['ab::zgType'.$zgId];
                 $zgPort = $config['ab::zgPort'.$zgId];
                 if (($zgType == "USB") || ($zgType == "USBv2")) {
+<<<<<<< HEAD
                     if ($config['ab::preventUsbPowerCycle'] == 'Y')
                         log::add('Abeille', 'Debug', 'Power cycle required for Zigate \'' . $zgId . '\' but disabled');
                     else {
@@ -498,6 +499,15 @@ class Abeille extends eqLogic {
                         log::add('Abeille', 'debug', 'Power cycling port \''.$zgPort.'\'');
                         exec($cmd." &"); // Exec in background
                     }
+=======
+if ($config['ab::preventUsbPowerCycle'] == 'Y') 
+                    log::add('Abeille', 'Debug', 'Power cycle required but disabled on \'' . $zgId . '\''); }
+
+                    $dir = __DIR__."/../scripts";
+                    $cmd = "cd ".$dir."; ".system::getCmdSudo()." ./powerCycleUsb.sh ".$zgPort." 1>/tmp/jeedom/Abeille/powerCycleUsb.log 2>&1";
+                    log::add('Abeille', 'debug', 'Power cycling port \''.$zgPort.'\'');
+                    exec($cmd." &"); // Exec in background
+>>>>>>> 3becde2c (Advanced option to disable dead USB Zigate power cycle)
                 } else if (($zgType == "PI") || ($zgType == "PIv2"))
                     Abeille::msgToCmd(PRIO_NORM, "CmdAbeille".$zgId."/0000/resetZg");
             }